[fix] Force the "Export spreadsheet" dialog to pop up on the active Window
On windows with multiple screens :
Open a first time the export spreadheet dialog on one of your screen (then close it).
Then if you move the main window to the second screen, and relaunch the export dialog,
the dialog will pop up on the previous screen.

This commit fix that.